requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==5.3.0
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.4.0
appnope==0.1.3
asgiref==3.5.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autopep8==1.6.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==1.2.2
black==21.11b1
bleach==6.1.0
blinker==1.8.1
bokeh==3.4.1
boto==2.49.0
boto3==1.24.3
botocore==1.27.3
cached-property==1.5.2
cachetools==5.3.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
comm==0.2.2
contourpy==1.2.1
cryptography==37.0.2
crytic-compile==0.2.2
cycler==0.12.1
Cython==3.0.10
cytoolz==0.11.2
dataclassy==0.11.1
decorator==5.1.1
distro==1.9.0
dj-database-url==0.5.0
Django==4.0.2
django-storages==1.12.3
dnspython==2.6.1
eip712==0.1.0
elevenlabs==0.2.26
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
exceptiongroup==1.1.3
execnet==1.9.0
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.2.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==20.1.0
h11==0.13.0
hexbytes==0.2.2
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.23.4
hume==0.3.3
hypothesis==6.27.3
idna==3.3
importlib-metadata==4.8.2
importlib_resources==6.4.0
inflection==0.5.0
iniconfig==1.1.1
insta-scrape==2.1.2
ipfshttpclient==0.8.0a2
ipydatawidgets==4.3.5
ipython==8.17.2
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.3
jiter==0.4.2
jmespath==1.0.0
joblib==1.4.2
jsonschema==3.2.0
jupyterlab_widgets==3.0.11
kittycad==0.6.14
kiwisolver==1.4.5
lapx==0.5.8
lazy-object-proxy==1.6.0
linkify-it-py==2.0.3
lru-dict==1.1.7
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
networkx==3.2.1
nltk==3.6.7
numpy==1.26.4
numpy-stl==3.1.1
openai==1.35.3
opencv-python==4.9.0.80
outcome==1.1.0
packaging==21.3
pandas==2.2.2
panel==1.4.4
param==2.1.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.0.0
pooch==1.8.2
prettytable==3.1.1
prompt-toolkit==3.0.40
protobuf==4.25.3
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyarrow==16.0.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic_core==2.18.4
pydeck==0.9.0
Pygments==2.18.0
pygments-lexer-solidity==0.7.0
PyGObject==3.42.0
PyJWT==1.7.1
pymongo==4.7.3
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pysha3==1.0.2
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.9.0.post0
python-dotenv==0.16.0
python-utils==3.8.2
pythreejs==2.4.2
pythx==1.6.1
pytube==15.0.0
pytz==2024.1
pyvista==0.43.10
pyviz_comms==3.0.2
PyYAML==5.4.1
regex==2021.11.10
requests==2.31.0
rich==13.7.1
rlp==2.0.1
s3transfer==0.6.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
scooby==0.10.0
seaborn==0.13.2
selenium==4.1.5
semantic-version==2.8.5
six==1.16.0
slither-analyzer==0.8.2
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.3
stpyvista==0.0.18
streamlit==1.34.0
streamlit-option-menu==0.3.13
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tk==0.1.0
tokenizers==0.19.1
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.13.0
traittypes==0.2.1
trame==3.6.3
trame-client==3.2.0
trame-server==3.0.2
trame-vtk==2.8.9
transformers==4.41.2
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.5.1
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
ultralytics==8.2.8
urllib3==1.26.7
varint==1.0.2
vtk==9.3.0
vvm==0.1.0
vyper==0.3.1
watchdog==4.0.1
wcwidth==0.2.5
web3==5.25.0
webencodings==0.5.1
websockets==11.0.3
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.13.3
wslink==2.1.1
wsproto==1.1.0
xyzservices==2024.6.0
yarl==1.7.2
zipp==3.6.0